Fix the resources directory issue
